big change

+++ song-store-assignment-1/song-store-assignment-1/src/utils/Utilities.java
@@ -22,8 +22,8 @@ public class Utilities {
     public static boolean YNtoBoolean(char charToSetVerified){
         return ((charToSetVerified == 'y') || (charToSetVerified == 'Y'));
     }
-    public static double toTwoDecimalPlaces(double number){
-        return (int) (number * 100 ) / 100.0;
+    public static int toTwoDecimalPlaces(int number){
+        return (int) (number * 100 ) / 100;
     }
 }
